+++ __tests/wow_api/wow_item_api.lua
@@ -43475,8 +43475,8 @@ _G.Items = {
 		itemEquipLoc = "INVTYPE_NON_EQUIP_IGNORE",
 		itemRarity = 4,
 		itemID = 236960,
-		itemClassId = 15,
-		itemSubClassId = 5,
+		itemClassID = 15,
+		itemSubClassID = 5,
 		itemSubType = "Mount",
 		itemType = "Miscellaneous",
 		itemString = "item:236960::::::::80:253::3::1:28:872"
@@ -43489,11 +43489,25 @@ _G.Items = {
 		itemEquipLoc = "INVTYPE_CHEST",
 		itemRarity = 4,
 		itemID = 228864,
-		itemClassId = 4,
-		itemSubClassId = 4,
+		itemClassID = 4,
+		itemSubClassID = 4,
 		itemSubType = "Leather",
 		itemType = "Armor",
 		itemString = "item:228864::::::::80:262::3:6:6652:11966:10354:11980:1494:10255:1:28:2462"
+	},
+	["item:264333"] = {
+		itemName = "Amani Incense Burner",
+		bindType = 1,
+		itemSellPrice = 0,
+		itemLink = "|cffa335ee|Hitem:264333:::::::::::|h[Amani Incense Burner]|h|r",
+		itemEquipLoc = "INVTYPE_NON_EQUIP_IGNORE",
+		itemRarity = 2,
+		itemID = 264333,
+		itemClassID = 20,
+		itemSubClassID = 0,
+		itemSubType = "Decor",
+		itemType = "Housing",
+		itemString = "item:264333:::::::::::",
 	}
 }